refactor(evm): isolate Monad system replay
Keep Monad protocol envelope parsing, prestate handling, and rollback in the feature-gated Monad module. Use an optional system replay hook so ordinary transactions retain generic execution and TxEnv mutation propagation.
