+++ .changelog/lean-monad-system-replay.md
@@ -1,0 +1,9 @@
+---
+anvil: patch
+cast: patch
+forge-verify: patch
+foundry-evm: patch
+foundry-evm-core: patch
+---
+
+Kept Monad protocol-system replay isolated while preserving ordinary system transaction handling.
