+++ crates/cast/src/cmd/run.rs
@@ -376,20 +376,8 @@ impl RunArgs {
 
         let factory = FEN::EvmFactory::default();
         let target_tx_env = TxEnvFor::<FEN>::from_recovered_tx(tx.as_ref(), tx.from());
-        let target_is_protocol_system = factory.protocol_system_call(&target_tx_env)?.is_some();
-
-        // Generic system transactions remain opt-in. Protocol system envelopes are always
-        // replayed through their network's dedicated execution path.
-        if !target_is_protocol_system
-            && !self.replay_system_txes
-            && (is_known_system_sender(tx.from())
-                || tx.transaction_type() == Some(SYSTEM_TRANSACTION_TYPE))
-        {
-            return Err(eyre::eyre!(
-                "{:?} is a system transaction.\nReplaying system transactions is currently not supported.",
-                tx.tx_hash()
-            ));
-        }
+        let target_is_system = is_known_system_sender(tx.from())
+            || tx.transaction_type() == Some(SYSTEM_TRANSACTION_TYPE);
 
         let tx_block_number = tx
             .block_number()
@@ -534,18 +522,8 @@ impl RunArgs {
                     }
 
                     let tx_env = TxEnvFor::<FEN>::from_recovered_tx(tx.as_ref(), tx.from());
-                    let is_protocol_system = factory.protocol_system_call(&tx_env)?.is_some();
-                    // Generic system transactions remain opt-in because they may omit pricing
-                    // fields. Protocol envelopes must be replayed to reconstruct canonical state.
-                    if !is_protocol_system
-                        && !self.replay_system_txes
-                        && (is_known_system_sender(tx.from())
-                            || tx.transaction_type() == Some(SYSTEM_TRANSACTION_TYPE))
-                    {
-                        pb.set_position((index + 1) as u64);
-                        continue;
-                    }
-
+                    let is_system = is_known_system_sender(tx.from())
+                        || tx.transaction_type() == Some(SYSTEM_TRANSACTION_TYPE);
                     let chain_context = block_context.as_ref().map_or_else(
                         || factory.chain_context_for_transaction(&tx_env),
                         |context| context.transaction(index),
@@ -553,22 +531,34 @@ impl RunArgs {
 
                     evm_env.cfg_env.disable_balance_check = true;
 
-                    if is_protocol_system {
-                        trace!(tx=?tx.tx_hash(), "executing previous protocol system transaction");
-                        executor
-                            .transact_protocol_system_with_env_and_context(
+                    if is_system {
+                        #[cfg(feature = "monad")]
+                        if executor
+                            .try_transact_system_replay_with_env_and_context(
                                 evm_env.clone(),
                                 tx_env.clone(),
-                                chain_context,
+                                chain_context.clone(),
                             )
                             .wrap_err_with(|| {
                                 format!(
-                                    "Failed to execute protocol system transaction: {:?} in block {}",
+                                    "Failed to replay system transaction: {:?} in block {}",
                                     tx.tx_hash(),
                                     evm_env.block_env.number()
                                 )
-                            })?;
-                    } else if let Some(to) = Transaction::to(tx) {
+                            })?
+                            .is_some()
+                        {
+                            trace!(tx=?tx.tx_hash(), "executed previous canonical system transaction");
+                            pb.set_position((index + 1) as u64);
+                            continue;
+                        }
+                        if !self.replay_system_txes {
+                            pb.set_position((index + 1) as u64);
+                            continue;
+                        }
+                    }
+
+                    if let Some(to) = Transaction::to(tx) {
                         trace!(tx=?tx.tx_hash(),?to, "executing previous call transaction");
                         executor
                             .transact_with_env_and_context(
@@ -651,28 +641,46 @@ impl RunArgs {
                 evm_env.cfg_env.disable_balance_check = true;
             }
 
-            if target_is_protocol_system {
-                trace!(tx=?tx.tx_hash(), "executing protocol system transaction");
-                TraceResult::from(executor.transact_protocol_system_with_env_and_context(
-                    evm_env,
-                    tx_env,
-                    chain_context,
-                )?)
-            } else if let Some(to) = Transaction::to(&tx) {
-                trace!(tx=?tx.tx_hash(), to=?to, "executing call transaction");
-                TraceResult::from(executor.transact_with_env_and_context(
-                    evm_env,
-                    tx_env,
-                    chain_context,
-                )?)
+            #[cfg(feature = "monad")]
+            let replay_result = if target_is_system {
+                executor.try_transact_system_replay_with_env_and_context(
+                    evm_env.clone(),
+                    tx_env.clone(),
+                    chain_context.clone(),
+                )?
             } else {
-                trace!(tx=?tx.tx_hash(), "executing create transaction");
-                TraceResult::try_from(executor.deploy_with_env_and_context(
-                    evm_env,
-                    tx_env,
-                    chain_context,
-                    None,
-                ))?
+                None
+            };
+            #[cfg(not(feature = "monad"))]
+            let replay_result: Option<foundry_evm::executors::RawCallResult<FEN>> = None;
+
+            if let Some(result) = replay_result {
+                trace!(tx=?tx.tx_hash(), "executed canonical system transaction");
+                TraceResult::from(result)
+            } else {
+                if target_is_system && !self.replay_system_txes {
+                    return Err(eyre::eyre!(
+                        "{:?} is a system transaction.\nReplaying system transactions is currently not supported.",
+                        tx.tx_hash()
+                    ));
+                }
+
+                if let Some(to) = Transaction::to(&tx) {
+                    trace!(tx=?tx.tx_hash(), to=?to, "executing call transaction");
+                    TraceResult::from(executor.transact_with_env_and_context(
+                        evm_env,
+                        tx_env,
+                        chain_context,
+                    )?)
+                } else {
+                    trace!(tx=?tx.tx_hash(), "executing create transaction");
+                    TraceResult::try_from(executor.deploy_with_env_and_context(
+                        evm_env,
+                        tx_env,
+                        chain_context,
+                        None,
+                    ))?
+                }
             }
         };
